feat: stable drip lists view

+++ src/lib/components/drip-list-card/drip-list-card-representational.svelte
@@ -26,6 +26,8 @@
   import type getIncomingSplitTotal from '$lib/utils/splits/get-incoming-split-total';
   import type { DripList } from '$lib/utils/metadata/types';
   import ChevronRight from 'radicle-design-system/icons/ChevronRight.svelte';
+  import TransitionedHeight from '../transitioned-height/transitioned-height.svelte';
+  import Spinner from '../spinner/spinner.svelte';
 
   export let dripList: DripList;
   export let format: 'thumblink' | 'full' = 'full';
@@ -142,7 +144,7 @@
 <svelte:element
   this={format === 'thumblink' ? 'a' : 'section'}
   href={format === 'thumblink' ? dripListUrl : undefined}
-  class="rounded-drip-lg shadow-low group transform {format === 'thumblink'
+  class="rounded-drip-lg overflow-hidden shadow-low group transform {format === 'thumblink'
     ? 'transition duration-200 mouse:hover:shadow-md mouse:hover:-translate-y-2px focus-visible:shadow-md focus-visible:-translate-y-2px'
     : ''}"
 >
@@ -172,7 +174,7 @@
           </div>
         {/if}
       </div>
-      {#if (dripList.description ?? '').length > 0}
+      {#if format === 'full' && (dripList.description ?? '').length > 0}
         <TextExpandable>
           {dripList.description}
         </TextExpandable>
@@ -186,27 +188,35 @@
         </div>
         <div class="typo-text tabular-nums total-streamed-badge">
           {#if browser}
-            <!-- TODO: Include incoming splits -->
             <AggregateFiatEstimate amounts={totalIncomingAmounts} />
           {/if}
           <span class="muted">&nbsp;total</span>
         </div>
-        {#if supportersPile && supportersPile.length > 0}
+        {#if supportersPile && supportersPile.length > 0 && format === 'full'}
           <div in:fade|local={{ duration: 300 }} class="supporters min-w-0">
             <span class="truncate muted">Supported by</span>
             <Pile components={supportersPile ?? []} itemsClickable={true} />
           </div>
         {/if}
       </div>
-      {#if representationalSplits}
-        <div class="splits-component">
-          <Splits list={representationalSplits} maxRows={maxSplitsRows} />
-        </div>
-      {:else}
-        Loading...
-      {/if}
+      <TransitionedHeight transitionHeightChanges={true}>
+        {#if representationalSplits}
+          <div in:fade class="splits-component">
+            <Splits
+              groupsExpandable={format === 'full'}
+              list={representationalSplits}
+              maxRows={maxSplitsRows}
+            />
+          </div>
+        {:else}
+          <div class="loading-state">
+            <Spinner />
+          </div>
+        {/if}
+      </TransitionedHeight>
     </div>
   </div>
+  <div class="overflow-gradient" />
 </svelte:element>
 
 <style>
@@ -214,6 +224,18 @@
     display: flex;
     align-items: center;
     gap: 1rem;
+    position: relative;
+  }
+
+  .overflow-gradient {
+    position: absolute;
+    z-index: 1;
+    width: 2rem;
+    top: 0;
+    right: 0;
+    bottom: 0;
+    /* gradient from transparent to var(--color-background), left to right */
+    background: linear-gradient(to right, transparent, var(--color-background) 75%);
   }
 
   .totals .drip-icon {
@@ -247,6 +269,13 @@
     display: flex;
     align-items: center;
     gap: 0.5rem;
+  }
+
+  .loading-state {
+    min-height: 212px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
   }
 
   .muted {
